backend/requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
dj-database-url==2.1.0
whitenoise==6.6.0
redis>=4.3.4
django-redis>=5.2.0
drf-yasg>=1.21.10
django-environ>=0.9.0
pytest>=7.4.0
pytest-django>=4.7.0
markdown>=3.5.2
coreapi>=2.3.3
django-ratelimit==4.1.0
flake8==6.1.0
flake8-docstrings==1.7.0
flake8-quotes==3.4.0
flake8-bugbear==22.10.27
flake8-comprehensions==3.14.0
flake8-import-order==0.18.2
pre-commit==3.6.0
black==23.12.1
isort==5.13.2